Update android package name and remove unnecessary code.
Remove and update unneeded files.

Update Android readme.
