+++ Android/README.md
@@ -1,17 +1,11 @@
-# Augmented Faces ARCore tutorials
-Augmented Faces for Android
+### Setup instructions for MMIW AR app.
+• You will need a real device for this app.</br>
+• Open Android Studio and click on `Open Recent Project`.</br>
+• Go to `/MMIWG2S/Android/` and open `settings.gradle`.</br>
+• This should open the project and gradle should sync.</br>
+• Choose Run in the task bar.
 
-<b> Try-On makeup with Augmented Faces </b> <br/>
-Blog post -> https://creativetech.blog/home/augmented-faces-makeup
-<br/>
-<br/>
-<b> Try-On glasses with Augmented Faces </b> <br/>
-Blog post -> https://creativetech.blog/home/try-on-glasses-arcore-augmented-faces
-<br/>
-<br/>
-<b> Instagram - like filter "Which ... are you?" with Augmented Faces </b> <br/>
-Blog post -> https://creativetech.blog/home/which-are-you-instagram-like-filter-arcore
-<br/>
-<br/>
-<b> Face Landmarks for ARCore Augmented Faces </b> <br/>
-Blog post -> https://creativetech.blog/home/face-landmarks-for-arcore-augmented-faces
+#### Reference
+This is the original tutorial used to start this project.</br>
+Blog post -> [https://creativetech.blog/home/augmented-faces-makeup](https://creativetech.blog/home/augmented-faces-makeup)</br>
+Source code -> [https://github.com/Kristina-Simakova/arfaces](https://github.com/Kristina-Simakova/arfaces)
