+++ Android/app/build.gradle
@@ -7,7 +7,7 @@ apply plugin: 'kotlin-android-extensions'
 android {
     compileSdkVersion 29
     defaultConfig {
-        applicationId "no.realitylab.arface"
+        applicationId "com.mehequanna.mmiw"
         minSdkVersion 24
         targetSdkVersion 29
         versionCode 1
@@ -31,20 +31,10 @@ dependencies {
     implementation"org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation 'androidx.appcompat:appcompat:1.1.0'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
-    implementation "com.google.ar.sceneform.ux:sceneform-ux:1.14.0"
+    implementation "com.google.ar.sceneform.ux:sceneform-ux:1.15.0"
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
 }
 
 apply plugin: 'com.google.ar.sceneform.plugin'
-
-sceneform.asset('sampledata/sunglasses.fbx',
-        'default',
-        'sampledata/sunglasses.sfa',
-        'src/main/assets/sunglasses')
-
-sceneform.asset('sampledata/yellow_sunglasses.fbx',
-        'default',
-        'sampledata/yellow_sunglasses.sfa',
-        'src/main/assets/yellow_sunglasses')
