--- Android/app/sampledata/sunglasses.sfa
@@ -1,60 +1,0 @@
-{
-  animations: [
-    {
-      path: 'sampledata/sunglasses.fbx',
-    },
-  ],
-  materials: [
-    {
-      name: '1A1A1A.001',
-      parameters: [
-        {
-          baseColor: [
-            0.101961,
-            0.101961,
-            0.101961,
-            1,
-          ],
-        },
-        {
-          baseColorMap: null,
-        },
-        {
-          normalMap: null,
-        },
-        {
-          interpolatedColor: null,
-        },
-        {
-          metallic: 0,
-        },
-        {
-          metallicMap: null,
-        },
-        {
-          roughness: 1,
-        },
-        {
-          roughnessMap: null,
-        },
-        {
-          opacity: null,
-        },
-      ],
-      source: 'build/sceneform_sdk/default_materials/fbx_material.sfm',
-    },
-  ],
-  model: {
-    attributes: [
-      'Position',
-      'Orientation',
-      'BoneIndices',
-      'BoneWeights',
-    ],
-    collision: {},
-    file: 'sampledata/sunglasses.fbx',
-    name: 'sunglasses',
-    recenter: 'root',
-  },
-  version: '0.54:2',
-}
