--- Android/app/sampledata/yellow_sunglasses.sfa
@@ -1,61 +1,0 @@
-{
-  animations: [
-    {
-      path: 'sampledata/yellow_sunglasses.fbx',
-    },
-  ],
-  materials: [
-    {
-      name: 'body',
-      parameters: [
-        {
-          baseColor: [
-            0.98101300000000002,
-            0.67214600000000002,
-            0,
-            1,
-          ],
-        },
-        {
-          baseColorMap: null,
-        },
-        {
-          normalMap: null,
-        },
-        {
-          interpolatedColor: null,
-        },
-        {
-          metallic: 0,
-        },
-        {
-          metallicMap: null,
-        },
-        {
-          roughness: 1,
-        },
-        {
-          roughnessMap: null,
-        },
-        {
-          opacity: null,
-        },
-      ],
-      source: 'build/sceneform_sdk/default_materials/fbx_material.sfm',
-    },
-  ],
-  model: {
-    attributes: [
-      'Position',
-      'TexCoord',
-      'Orientation',
-      'BoneIndices',
-      'BoneWeights',
-    ],
-    collision: {},
-    file: 'sampledata/yellow_sunglasses.fbx',
-    name: 'yellow_sunglasses',
-    recenter: 'root',
-  },
-  version: '0.54:2',
-}
