+++ Android/app/src/androidTest/java/com/mehequanna/mmiw/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package no.realitylab.arface
+package com.mehequanna.mmiw
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getTargetContext()
-        assertEquals("no.realitylab.arface", appContext.packageName)
+        assertEquals("com.mehequanna.mmiw", appContext.packageName)
     }
 }
