+++ Android/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
           xmlns:tools="http://schemas.android.com/tools"
-          package="no.realitylab.arface">
+          package="com.mehequanna.mmiw">
 
     <uses-permission android:name="android.permission.CAMERA" />
     <!-- This tag indicates that this application requires ARCore. This results in the application
@@ -19,7 +19,7 @@
             android:theme="@style/AppTheme">
 
         <meta-data android:name="com.google.ar.core" android:value="required" />
-        <activity android:name=".MainActivity"
+        <activity android:name="com.mehequanna.mmiw.MainActivity"
                   android:configChanges="orientation|screenSize"
                   android:theme="@style/Theme.AppCompat.NoActionBar"
                   android:screenOrientation="locked"
@@ -30,23 +30,10 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity android:name=".MakeupActivity"
+        <activity android:name="com.mehequanna.mmiw.MmiwActivity"
             android:configChanges="orientation|screenSize"
             android:theme="@style/Theme.AppCompat.NoActionBar"
             android:screenOrientation="locked"/>
-        <activity android:name=".GlassesActivity"
-            android:configChanges="orientation|screenSize"
-            android:theme="@style/Theme.AppCompat.NoActionBar"
-            android:screenOrientation="locked"/>
-        <activity android:name=".FaceRegionsActivity"
-            android:configChanges="orientation|screenSize"
-            android:theme="@style/Theme.AppCompat.NoActionBar"
-            android:screenOrientation="locked"/>
-        <activity android:name=".FaceLandmarksActivity"
-            android:configChanges="orientation|screenSize"
-            android:theme="@style/Theme.AppCompat.NoActionBar"
-            android:screenOrientation="locked"/>
-
     </application>
 
 </manifest>
