+++ Android/app/src/main/java/com/mehequanna/mmiw/FaceArFragment.kt
@@ -1,4 +1,4 @@
-package no.realitylab.arface
+package com.mehequanna.mmiw
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -10,7 +10,7 @@ import com.google.ar.core.Session
 import com.google.ar.sceneform.ux.ArFragment
 import java.util.*
 
-public class FaceArFragment : ArFragment() {
+class FaceArFragment : ArFragment() {
     override fun getSessionConfiguration(session: Session?): Config {
         val config = Config(session)
         config.augmentedFaceMode = Config.AugmentedFaceMode.MESH3D
