+++ Android/app/src/main/java/com/mehequanna/mmiw/MainActivity.kt
@@ -1,0 +1,19 @@
+package com.mehequanna.mmiw
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import kotlinx.android.synthetic.main.activity_main.*
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        setContentView(R.layout.activity_main)
+        button_makeup.setOnClickListener {
+            startActivity(Intent(this, MmiwActivity::class.java))
+        }
+    }
+
+}
