+++ Android/app/src/main/java/com/mehequanna/mmiw/MmiwActivity.kt
@@ -1,35 +1,26 @@
-package no.realitylab.arface
+package com.mehequanna.mmiw
 
 import android.app.ActivityManager
 import android.content.Context
-import android.net.Uri
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.google.ar.core.ArCoreApk
 import com.google.ar.core.AugmentedFace
 import com.google.ar.core.TrackingState
-import com.google.ar.sceneform.rendering.ModelRenderable
 import com.google.ar.sceneform.rendering.Renderable
 import com.google.ar.sceneform.rendering.Texture
 import com.google.ar.sceneform.ux.AugmentedFaceNode
-import kotlinx.android.synthetic.main.activity_glasses.*
-import java.util.ArrayList
+import kotlinx.android.synthetic.main.activity_makeup.*
 
-class GlassesActivity : AppCompatActivity() {
-
+class MmiwActivity : AppCompatActivity() {
     companion object {
         const val MIN_OPENGL_VERSION = 3.0
     }
 
     lateinit var arFragment: FaceArFragment
     private var faceMeshTexture: Texture? = null
-    private var glasses: ArrayList<ModelRenderable> = ArrayList()
-    private var faceRegionsRenderable: ModelRenderable? = null
-
     var faceNodeMap = HashMap<AugmentedFace, AugmentedFaceNode>()
-    private var index: Int = 0
-    private var changeModel: Boolean = false
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -37,60 +28,29 @@ class GlassesActivity : AppCompatActivity() {
             return
         }
 
-        setContentView(R.layout.activity_glasses)
-        button_next.setOnClickListener {
-            changeModel = !changeModel
-            index++
-            if (index > glasses.size - 1) {
-                index = 0
-            }
-            faceRegionsRenderable = glasses.get(index)
-        }
-
+        setContentView(R.layout.activity_makeup)
         arFragment = face_fragment as FaceArFragment
         Texture.builder()
-            .setSource(this, R.drawable.makeup)
+            .setSource(this, R.drawable.red_hand_80_texture)
             .build()
             .thenAccept { texture -> faceMeshTexture = texture }
-
-        ModelRenderable.builder()
-            .setSource(this, Uri.parse("yellow_sunglasses.sfb"))
-            .build()
-            .thenAccept { modelRenderable ->
-                glasses.add(modelRenderable)
-                faceRegionsRenderable = modelRenderable
-                modelRenderable.isShadowCaster = false
-                modelRenderable.isShadowReceiver = false
-            }
-
-        ModelRenderable.builder()
-            .setSource(this, Uri.parse("sunglasses.sfb"))
-            .build()
-            .thenAccept { modelRenderable ->
-                glasses.add(modelRenderable)
-                modelRenderable.isShadowCaster = false
-                modelRenderable.isShadowReceiver = false
-            }
 
         val sceneView = arFragment.arSceneView
         sceneView.cameraStreamRenderPriority = Renderable.RENDER_PRIORITY_FIRST
         val scene = sceneView.scene
 
         scene.addOnUpdateListener {
-            if (faceRegionsRenderable != null) {
+            faceMeshTexture.let {
                 sceneView.session
                     ?.getAllTrackables(AugmentedFace::class.java)?.let {
                         for (f in it) {
                             if (!faceNodeMap.containsKey(f)) {
                                 val faceNode = AugmentedFaceNode(f)
                                 faceNode.setParent(scene)
-                                faceNode.faceRegionsRenderable = faceRegionsRenderable
+                                faceNode.faceMeshTexture = faceMeshTexture
                                 faceNodeMap.put(f, faceNode)
-                            } else if (changeModel) {
-                                faceNodeMap.getValue(f).faceRegionsRenderable = faceRegionsRenderable
                             }
                         }
-                        changeModel = false
                         // Remove any AugmentedFaceNodes associated with an AugmentedFace that stopped tracking.
                         val iter = faceNodeMap.entries.iterator()
                         while (iter.hasNext()) {
