--- Android/app/src/main/java/no/realitylab/arface/FaceLandmarksActivity.kt
@@ -1,83 +1,0 @@
-package no.realitylab.arface
-
-import android.app.ActivityManager
-import android.content.Context
-import android.os.Bundle
-import android.view.View
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import com.google.ar.core.ArCoreApk
-import com.google.ar.core.AugmentedFace
-import com.google.ar.core.TrackingState
-import com.google.ar.sceneform.rendering.Renderable
-import kotlinx.android.synthetic.main.activity_regions.*
-
-class FaceLandmarksActivity : AppCompatActivity() {
-    companion object {
-        const val MIN_OPENGL_VERSION = 3.0
-    }
-
-    lateinit var arFragment: FaceArFragment
-    var faceNodeMap = HashMap<AugmentedFace, CustomFaceNode>()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        if (!checkIsSupportedDeviceOrFinish()) {
-            return
-        }
-
-        setContentView(R.layout.activity_regions)
-        arFragment = face_fragment as FaceArFragment
-
-        button_refresh.visibility = View.GONE
-
-        val sceneView = arFragment.arSceneView
-        sceneView.cameraStreamRenderPriority = Renderable.RENDER_PRIORITY_FIRST
-        val scene = sceneView.scene
-
-        scene.addOnUpdateListener {
-            sceneView.session
-                ?.getAllTrackables(AugmentedFace::class.java)?.let {
-                    for (f in it) {
-                        if (!faceNodeMap.containsKey(f)) {
-                            val faceNode = CustomFaceNode(f, this)
-                            faceNode.setParent(scene)
-                            faceNodeMap.put(f, faceNode)
-                        }
-                    }
-                    // Remove any AugmentedFaceNodes associated with an AugmentedFace that stopped tracking.
-                    val iter = faceNodeMap.entries.iterator()
-                    while (iter.hasNext()) {
-                        val entry = iter.next()
-                        val face = entry.key
-                        if (face.trackingState == TrackingState.STOPPED) {
-                            val faceNode = entry.value
-                            faceNode.setParent(null)
-                            iter.remove()
-                        }
-                    }
-                }
-        }
-    }
-
-    private fun checkIsSupportedDeviceOrFinish() : Boolean {
-        if (ArCoreApk.getInstance().checkAvailability(this) == ArCoreApk.Availability.UNSUPPORTED_DEVICE_NOT_CAPABLE) {
-            Toast.makeText(this, "Augmented Faces requires ARCore", Toast.LENGTH_LONG).show()
-            finish()
-            return false
-        }
-        val openGlVersionString =  (getSystemService(Context.ACTIVITY_SERVICE) as? ActivityManager)
-            ?.deviceConfigurationInfo
-            ?.glEsVersion
-
-        openGlVersionString?.let { s ->
-            if (java.lang.Double.parseDouble(openGlVersionString) < MIN_OPENGL_VERSION) {
-                Toast.makeText(this, "Sceneform requires OpenGL ES 3.0 or later", Toast.LENGTH_LONG)
-                    .show()
-                finish()
-                return false
-            }
-        }
-        return true
-    }
-}
