+++ Android/app/src/main/res/layout/activity_main.xml
@@ -12,22 +12,4 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/makeup" />
-
-    <Button
-        android:id="@+id/button_glasses"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/glasses" />
-
-    <Button
-        android:id="@+id/button_regions"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/regions" />
-
-    <Button
-        android:id="@+id/button_face_landmarks"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/face_landmarks" />
 </LinearLayout>
