--- Android/app/src/main/res/layout/card_layout.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="horizontal"
-    android:layout_width="wrap_content"
-    android:background="@drawable/rounded_bg"
-    android:layout_height="wrap_content"
-    android:clipChildren="false"
-    android:padding="8dp"
-    android:gravity="center"
-    android:clipToPadding="false">
-
-    <TextView
-        android:id="@+id/title"
-        android:layout_width="50dp"
-        android:layout_height="wrap_content"
-        android:textAlignment="center"
-        android:textSize="8sp"
-        android:text="@string/quiz_title"
-        android:padding="8dp"/>
-
-</LinearLayout>
