--- Android/app/src/main/res/layout/element_layout.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="horizontal"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:clipChildren="false"
-    android:gravity="center"
-    android:clipToPadding="false">
-
-    <ImageView
-        android:id="@+id/element_image"
-        android:src="@drawable/star_eye"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content" />
-
-</LinearLayout>
