+++ Android/app/src/main/res/values/strings.xml
@@ -1,8 +1,4 @@
 <resources>
-    <string name="app_name">arface</string>
-    <string name="makeup">Makeup tutorial</string>
-    <string name="glasses">Glasses tutorial</string>
-    <string name="regions">Quiz tutorial</string>
-    <string name="face_landmarks">Face landmarks</string>
-    <string name="quiz_title">What animal are you?</string>
+    <string name="app_name">MMIW AR</string>
+    <string name="makeup">MMIW Prototype</string>
 </resources>
